Allow drag events to access files
Fixes a regression. When files are dragged into Chromium, it does not
allow JS to access the File objects until they're dropped. Earlier
versions of Electron (and Chromium/WebKit) supported this, but it broke
when WebKit was updated to comply with the specification.

This patch allows JS to access File objects before they're dropped. It's
limited to only local URLs (ones starting with file://) and to
localhost.